Travis...

+++ demo/15.file.download.spec.js
@@ -19,9 +19,11 @@ describe('Downloading a dynamically generated file', function() {
         else {
 
             var downloadFolder = path.join(process.cwd(), 'demo', 'download')
-            var files = fs.readdirSync(downloadFolder)
-            for (let i=0; i<files.length; i++) {
-                fs.unlinkSync(path.join(process.cwd(), 'demo', 'download', files[i]))
+            if (fs.existsSync(downloadFolder)) {
+                var files = fs.readdirSync(downloadFolder)
+                for (let i=0; i<files.length; i++) {
+                    fs.unlinkSync(path.join(process.cwd(), 'demo', 'download', files[i]))
+                }
             }
 
             let options = new firefox.Options()
